requirements.txt
openai
streamlit
python-dotenv
sounddevice
soundfile
numpy
scipy
